fix: remove circular dependencies from cli
