+++ packages/cli/src/cli.ts
@@ -82,7 +82,6 @@ import {
   DEFAULT_CONFIG_TESTNET_PATH,
   ID_ADDRESS_PATTERN,
   STACKS_ADDRESS_PATTERN,
-  DEFAULT_MAX_ID_SEARCH_INDEX,
 } from './argparse';
 
 import { encryptBackupPhrase, decryptBackupPhrase } from './encrypt';
@@ -93,7 +92,6 @@ import { gaiaAuth, gaiaConnect, gaiaUploadProfileAll, getGaiaAddressFromProfile 
 
 import {
   JSONStringify,
-  getPrivateKeyAddress,
   canonicalPrivateKey,
   decodePrivateKey,
   makeProfileJWT,
@@ -112,7 +110,7 @@ import {
 
 import { handleAuth, handleSignIn } from './auth';
 import { generateNewAccount, generateWallet, getAppPrivateKey } from '@stacks/wallet-sdk';
-
+import { getMaxIDSearchIndex, setMaxIDSearchIndex, getPrivateKeyAddress } from './common';
 // global CLI options
 let txOnly = false;
 let estimateOnly = false;
@@ -120,13 +118,8 @@ let safetyChecks = true;
 let receiveFeesPeriod = 52595;
 let gracePeriod = 5000;
 let noExit = false;
-let maxIDSearchIndex = DEFAULT_MAX_ID_SEARCH_INDEX;
 
 let BLOCKSTACK_TEST = !!process.env.BLOCKSTACK_TEST;
-
-export function getMaxIDSearchIndex() {
-  return maxIDSearchIndex;
-}
 
 /*
  * Sign a profile.
@@ -1804,8 +1797,10 @@ export function CLIMain() {
     safetyChecks = !CLIOptAsBool(opts, 'U');
     receiveFeesPeriod = opts['N'] ? parseInt(CLIOptAsString(opts, 'N')!) : receiveFeesPeriod;
     gracePeriod = opts['G'] ? parseInt(CLIOptAsString(opts, 'N')!) : gracePeriod;
-    maxIDSearchIndex = opts['M'] ? parseInt(CLIOptAsString(opts, 'M')!) : maxIDSearchIndex;
-
+    const maxIDSearchIndex = opts['M']
+      ? parseInt(CLIOptAsString(opts, 'M')!)
+      : getMaxIDSearchIndex();
+    setMaxIDSearchIndex(maxIDSearchIndex);
     const debug = CLIOptAsBool(opts, 'd');
     const consensusHash = CLIOptAsString(opts, 'C');
     const integration_test = CLIOptAsBool(opts, 'i');
